Removed unnecessary frame changes
keyframe_insert() takes the frame number that was passed to it; the current frame doesn't have to be set to that frame for it to work. This has the added advantage that the artist using this operator doesn't have to go back to the frame they were on before using it.