New cubemap scheme, new way to configure the name.
Sometimes it is easier to generate a cubmap by animating the camera,
instead of creating a new scene per view. That also helps when you are
using a renderfarm, since you can distribute the job more easily.
In order to generate a cubemap you first select the source folder. The
script will then figure out if the cubemaps where saved as scenes
("NORTH", "SOUTH", etc) or as animation ("000001", "000002", etc)
If you leave the cubemap name field empty, the script will use the name
of the source folder as name for the cubemap