Recorded actions are not assigned to anymore
Instead there is now a playback button for the virtual camera.
You can pick from all actions from the scene.
You can also delete the actions from there.
What is still missing though is a way to only use the animation during
playback, and as soon as playback is cancelt, return to previous location and unassign the action